% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxpower.R
\name{fxpower_alpha_star}
\alias{fxpower_alpha_star}
\title{Empirical Type I Error for Fisher's Exact Test}
\usage{
fxpower_alpha_star(n1, n2, p0, alpha, eps = 1e-06)
}
\arguments{
\item{n1}{Integer. Sample size in group 1.}

\item{n2}{Integer. Sample size in group 2.}

\item{p0}{Numeric. Common event probability under H0.}

\item{alpha}{Numeric. Nominal significance level.}

\item{eps}{Numeric. Trimming parameter.}
}
\value{
Numeric scalar: the empirical type I error rate.
}
\description{
Compute alpha* by evaluating the power function under H0
(p1 = p2 = p0).
}
\keyword{internal}
