% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{net_vander_cpp}
\alias{net_vander_cpp}
\title{Network enumeration with Vandermonde bulk-sum pruning (C++)}
\usage{
net_vander_cpp(dat)
}
\arguments{
\item{dat}{Integer matrix with 2 columns (r x 2 contingency table).}
}
\value{
An htest object matching \code{fisher.test()} output.
}
\description{
Network enumeration with Vandermonde bulk-sum pruning (C++)
}
