% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_memo.R
\name{tree_memo}
\alias{tree_memo}
\title{Memoized tree traversal (pure R)}
\usage{
tree_memo(dat)
}
\arguments{
\item{dat}{Integer matrix with 2 columns (r x 2 contingency table).}
}
\value{
An htest object matching \code{fisher.test()} output.
}
\description{
Memoized tree traversal (pure R)
}
