% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tree_memo_profile}
\alias{tree_memo_profile}
\title{Memoized tree traversal with profiling (C++)}
\usage{
tree_memo_profile(dat)
}
\arguments{
\item{dat}{Integer matrix with 2 columns (r x 2 contingency table).}
}
\value{
An htest object with a \code{$profile} list of counters.
}
\description{
Memoized tree traversal with profiling (C++)
}
