% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzfisher-package.R
\docType{package}
\name{zzfisher-package}
\alias{zzfisher}
\alias{zzfisher-package}
\title{zzfisher: Fisher's Exact Test for r x 2 Contingency Tables}
\description{
Multiple algorithmic implementations of Fisher's exact test for r x 2 contingency tables. Includes tree-based (mode-centric traversal with pruning) and network-based (state-space enumeration with Vandermonde bulk-sum) approaches in pure R, C, and C++.
}
\author{
\strong{Maintainer}: RG Thomas \email{rgthomas@ucsd.edu} (\href{https://orcid.org/0000-0003-1686-4965}{ORCID})

}
\keyword{internal}
